+++ src/bump_scope.rs
@@ -12,13 +12,14 @@ use core::{
 };
 
 use crate::{
-    BaseAllocator, BumpBox, BumpScopeGuard, BumpString, BumpVec, Checkpoint, ErrorBehavior, FixedBumpString, FixedBumpVec,
-    MinimumAlignment, MutBumpString, MutBumpVec, MutBumpVecRev, NoDrop, RawChunk, SizedTypeProperties,
+    BaseAllocator, BumpAllocator, BumpBox, BumpScopeGuard, BumpString, BumpVec, Checkpoint, ErrorBehavior, FixedBumpString,
+    FixedBumpVec, MinimumAlignment, MutBumpString, MutBumpVec, MutBumpVecRev, NoDrop, RawChunk, SizedTypeProperties,
     SupportedMinimumAlignment,
     alloc::{AllocError, Allocator},
     allocator_impl,
     bump_align_guard::BumpAlignGuard,
     bumping::{BumpUp, bump_down, bump_up},
+    chunk_header::unallocated_chunk_header,
     chunk_size::ChunkSize,
     const_param_assert, down_align_usize,
     layout::{ArrayLayout, CustomLayout, LayoutProps, SizedLayout},
@@ -331,7 +332,14 @@ where
     pub fn scope_guard(&mut self) -> BumpScopeGuard<'_, A, MIN_ALIGN, UP> {
         BumpScopeGuard::new(self)
     }
+}
 
+impl<'a, A, const MIN_ALIGN: usize, const UP: bool, const GUARANTEED_ALLOCATED: bool>
+    BumpScope<'a, A, MIN_ALIGN, UP, GUARANTEED_ALLOCATED>
+where
+    MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
+    A: BaseAllocator<GUARANTEED_ALLOCATED>,
+{
     /// Creates a checkpoint of the current bump position.
     ///
     /// The bump position can be reset to this checkpoint with [`reset_to`].
@@ -365,8 +373,11 @@ where
     /// # Safety
     ///
     /// - the checkpoint must have been created by this bump allocator
-    /// - the bump allocator must not have been [`reset`](crate::Bump::reset) since creation of this checkpoint
+    /// - the bump allocator must not have been [`reset`] since creation of this checkpoint
     /// - there must be no references to allocations made since creation of this checkpoint
+    /// - the checkpoint must not have been created by an`!GUARANTEED_ALLOCATED` when self is `GUARANTEED_ALLOCATED`
+    ///
+    /// [`reset`]: crate::Bump::reset
     ///
     /// # Examples
     ///
@@ -385,25 +396,55 @@ where
     /// assert_eq!(bump.stats().allocated(), 0);
     /// ```
     #[inline]
+    #[expect(clippy::missing_panics_doc)] // just debug assertions
     pub unsafe fn reset_to(&self, checkpoint: Checkpoint) {
-        debug_assert!(self.stats().big_to_small().any(|chunk| {
-            chunk.header == checkpoint.chunk.cast() && chunk.contains_addr_or_end(checkpoint.address.get())
-        }));
-
         unsafe {
-            checkpoint.reset_within_chunk();
-            let chunk = RawChunk::from_header(checkpoint.chunk.cast());
-            self.chunk.set(chunk);
+            // If the checkpoint was created when the bump allocator had no allocated chunk
+            // then the chunk pointer will point to the unallocated chunk header.
+            //
+            // In such cases we reset the bump pointer to the very start of the very first chunk.
+            //
+            // We don't check if the chunk pointer points to the unallocated chunk header
+            // if the bump allocator is `GUARANTEED_ALLOCATED`. We are allowed to not do this check
+            // because of this safety condition of `reset_to`:
+            // > the checkpoint must not have been created by an`!GUARANTEED_ALLOCATED` when self is `GUARANTEED_ALLOCATED`
+            if !GUARANTEED_ALLOCATED && checkpoint.chunk == unallocated_chunk_header() {
+                let mut chunk = self.chunk.get();
+
+                while let Some(prev) = chunk.prev() {
+                    chunk = prev;
+                }
+
+                chunk.reset();
+                self.chunk.set(chunk);
+            } else {
+                debug_assert_ne!(
+                    checkpoint.chunk,
+                    unallocated_chunk_header(),
+                    "the safety conditions state that \"the checkpoint must not have been created by an`!GUARANTEED_ALLOCATED` when self is `GUARANTEED_ALLOCATED`\""
+                );
+
+                #[cfg(debug_assertions)]
+                {
+                    let chunk = self
+                        .stats()
+                        .small_to_big()
+                        .find(|chunk| chunk.header == checkpoint.chunk.cast())
+                        .expect("this checkpoint does not refer to any chunk of this bump allocator");
+
+                    assert!(
+                        chunk.contains_addr_or_end(checkpoint.address.get()),
+                        "checkpoint address does not point within its chunk"
+                    );
+                }
+
+                checkpoint.reset_within_chunk();
+                let chunk = RawChunk::from_header(checkpoint.chunk.cast());
+                self.chunk.set(chunk);
+            }
         }
     }
-}
 
-impl<'a, A, const MIN_ALIGN: usize, const UP: bool, const GUARANTEED_ALLOCATED: bool>
-    BumpScope<'a, A, MIN_ALIGN, UP, GUARANTEED_ALLOCATED>
-where
-    MinimumAlignment<MIN_ALIGN>: SupportedMinimumAlignment,
-    A: BaseAllocator<GUARANTEED_ALLOCATED>,
-{
     #[inline(always)]
     pub(crate) unsafe fn new_unchecked(chunk: RawChunk<UP, A>) -> Self {
         Self {