+++ src/stats.rs
@@ -404,7 +404,7 @@ impl<'a, A, const UP: bool> Chunk<'a, A, UP> {
         self.raw().pos()
     }
 
-    #[inline(always)]
+    #[cfg(debug_assertions)]
     pub(crate) fn contains_addr_or_end(self, addr: usize) -> bool {
         self.raw().contains_addr_or_end(addr)
     }